update v1.7.50 build-6982

+++ FastPhotoTweet/Twitter/TWCharCounter.h
@@ -9,6 +9,6 @@
 
 @interface TWCharCounter : NSObject
 
-+ (int)charCounter:(id)post;
++ (NSInteger)charCounter:(id)post;
 
 @end
